Reformat project structure.

+++ src/main/java/com/emp/app/objects/dao/DBObjectDAO.java
@@ -9,6 +9,6 @@ public interface DBObjectDAO {
 	public void addDBObject(DBObject p);
 	public void updateDBObject(DBObject p);
 	public List<DBObject> listDBObjects();
-	public DBObject getDBObjectById(int id);
-	public void removeDBObject(int id);
+	public DBObject getDBObjectById(long id);
+	public void removeDBObject(long id);
 }
